fixed bug and finished notes
changed the password string to empty and changed variables to false inside the while loop and after the for loop.

This ensures if the if statement returns a false that we will no longer double the length of the password and the if statement is honest again by setting the necessary variables to false.